renamed buttonInteraction to handleButtons
